added grunt serve to the workflow
