+++ package.json
@@ -27,6 +27,7 @@
       "font-awesome":                 "^4.6.3",
       "body-parser":                  "~1.13.2",
       "cookie-parser":                "~1.3.5",
+      "connect-livereload":           "0.6.0",
       "debug":                        "~2.2.0",
       "express":                      "4.14.0",
       "pug":                          "2.0.0-beta6",
@@ -45,7 +46,11 @@
       "grunt-contrib-htmlmin":        "2.0.0",
       "grunt-contrib-symlink":        "1.0.0",
       "grunt-contrib-compress":       "1.3.0",
+      "grunt-contrib-watch":          "1.0.0",
       "grunt-jsdoc":                  "2.0.0",
+      "grunt-concurrent":             "2.3.1",
+      "grunt-nodemon":                "0.4.2",
+      "grunt-open":                   "0.2.3",
       "time-grunt":                   "1.3.0"
     }
 }
